chore(Storybook): Split AppIcons into per-component files under AppIcon
Give each app-icon showcase its own file — AppleTouchIcon.tsx,
Favicon.tsx, SvgIcon.tsx and WebAppIcons.tsx — and rename the directory
AppIcons → AppIcon to match the singular app-icon.css / _ams-app-icon
naming. The aggregate test-stories file follows the directory rename, and
the three importers (two docs pages and the test file) now pull each
component from its own file. Every file imports app-icon.css; the bundler
dedupes it.
