+++ storybook/src/brand/assets/favicon.docs.mdx
@@ -2,7 +2,9 @@
 
 import { Meta } from "@storybook/addon-docs/blocks";
 
-import { AppleTouchIcon, Favicon, SvgIcon } from "#storybook/_components/AppIcons/AppIcons";
+import { AppleTouchIcon } from "#storybook/_components/AppIcon/AppleTouchIcon";
+import { Favicon } from "#storybook/_components/AppIcon/Favicon";
+import { SvgIcon } from "#storybook/_components/AppIcon/SvgIcon";
 
 <Meta title="Brand/Assets/Favicon" />
