+++ storybook/src/docs/developer-guide/web-app.docs.mdx
@@ -2,7 +2,7 @@
 
 import { Meta } from "@storybook/addon-docs/blocks";
 
-import { WebAppIcons } from "#storybook/_components/AppIcons/AppIcons";
+import { WebAppIcons } from "#storybook/_components/AppIcon/WebAppIcons";
 
 <Meta title="Docs/Developer Guide/Web App" />
